Addressed comments on request. Removed MouseController and associated objects. Removed font references. Removed Mouse dictionary. Moved cycle commands to their own folder.
